don't kick with rx_rate higher threshold
