+++ files/dawn.config
@@ -31,3 +31,4 @@ config settings metric
     option rssi_val         '-60'
     option low_rssi_val     '-79'
     option min_probe_count  '2'
+    option bandwith_threshold '6'
