+++ src/CMakeLists.txt
@@ -39,7 +39,7 @@ SET(SOURCES
         crypto/base64.c
         include/base64.h
 
-        utils/utils.c include/rssi.h utils/rssi.c)
+        utils/utils.c include/dawn_iwinfo.h utils/dawn_iwinfo.c)
 
 SET(LIBS
         ubox ubus json-c blobmsg_json config uci gcrypt iwinfo)
